+++ CHANGELOG.md
@@ -7,6 +7,11 @@ project adheres to [Semantic Versioning](http://semver.org/).
 
 <!-- ## Unreleased -->
 
+### Added
+
+-   Added a `globalMeasurementExpression` option for overriding the default
+    global measurement of `window.tachometerResult` to an arbitrary expression.
+
 ## [0.4.13] 2019-09-12
 
 ### Added